Connection details for remoties
I put this up in Yammer and people are confused by the current instructions on how to get this up and running by themselves.  I hope this might help non webdev folks have an instance up in their home offices.  Took out the stuff about changing the location tag since that seems to be hit or miss on my end, it's very opaque and the average new user will not need it for now.  I'm happy to try and write more documentation for it to flesh out the intended usage.