advertise how to change content
I don't think people realize it is possible to define new content to put
on the screens in the offices. To spread this knowledge, we introduce a
new piece of content that tells people where the content lives.

The image is currently the infamous "johnath pointing" image with "Your
Content Here" (in Comic Sans of course) with a URL to the moz-corsica
GitHub project. And just to troll our UX people, the image features two
colors from Mozilla's brand guidelines that shouldn't be used in direct
contrast. This will surely invite others to add "proper" content.